+++ core/components/twofactorx/processors/mgr/email/instructions.class.php
@@ -35,7 +35,7 @@ class TwoFactorXEmailInstructionsProcessor extends Processor
             }
             return $this->modx->error->success($this->modx->lexicon('twofactorx.email_success'));
         } else {
-            return $this->modx->error->failure($this->modx->lexicon('twofactorx.invaliddata'));
+            return $this->modx->error->failure($this->modx->lexicon('twofactorx.error_invalid_data'));
         }
     }