+++ core/components/twofactorx/src/Plugins/Events/OnBeforeManagerLogin.php
@@ -24,26 +24,26 @@ class OnBeforeManagerLogin extends Plugin
 
         if ($this->twofactorx->getOption('enable_2fa')) {
             $this->modx->controller->addLexiconTopic('twofactorx:default');
-            if (!$this->twofactorx->userExist) {
+            if (!$this->twofactorx->getUserExist()) {
                 $output = true;
-            } else if ($this->twofactorx->userStatus) {
+            } else if ($this->twofactorx->getUserTotpDisabled()) {
                 $output = true;
                 if ($this->twofactorx->getOption('debug')) {
-                    $this->modx->log(xPDO::LOG_LEVEL_ERROR, "Authenticator for user $username is disabled.");
+                    $this->modx->log(xPDO::LOG_LEVEL_ERROR, "Two Factor Authentication for user $username is disabled.");
                 }
-            } else if ($this->twofactorx->getOption('enable_onetime') && $this->twofactorx->userOnetimeStatus) {
+            } else if ($this->twofactorx->getUserVerifyTotpStatus()) {
                 $output = true;
                 if ($this->twofactorx->getOption('debug')) {
-                    $this->modx->log(xPDO::LOG_LEVEL_ERROR, "TwoFactorX: user $username logged in onetime mode.");
+                    $this->modx->log(xPDO::LOG_LEVEL_ERROR, "User $username logged in verify TOTP mode.");
                 }
             } else if (empty($code)) {
                 $output = $this->modx->lexicon('twofactorx.enterkey');
             } else if (preg_match("/^[0-9]{6}$/", $code) < 1) {
-                $output = $this->modx->lexicon('twofactorx.invalidformat');
+                $output = $this->modx->lexicon('twofactorx.error_invalid_format');
             } else if ($this->twofactorx->userCodeMatch($code)) {
                 $output = true;
             } else {
-                $output = $this->modx->lexicon('twofactorx.invalidcode');
+                $output = $this->modx->lexicon('twofactorx.error_invalid_key');
             }
         } else {
             $output = true;